+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgentCoordinator.scala
@@ -168,7 +168,7 @@ object GridAgentCoordinator {
   }
 
   def apply(
-      config: SimonaConfig.Simona,
+      config: SimonaConfig,
       subgrids: Seq[SubGridContainer],
   )(using environmentRefs: EnvironmentRefs): Behavior[Message] =
     Behaviors.setup { ctx =>
@@ -437,7 +437,7 @@ object GridAgentCoordinator {
       subgrids: Seq[SubGridContainer],
       resolution: Long,
       pfParams: PowerFlowParams,
-      cfg: SimonaConfig.Simona,
+      cfg: SimonaConfig,
   )(using
       context: ActorContext[Message],
       environmentRefs: EnvironmentRefs,