+++ src/main/scala/edu/ie3/simona/agent/grid/powerflow/PowerFlowParams.scala
@@ -6,7 +6,7 @@
 
 package edu.ie3.simona.agent.grid.powerflow
 
-import edu.ie3.simona.config.SimonaConfig.Simona.Powerflow
+import edu.ie3.simona.config.SimonaConfig.Powerflow
 
 import scala.concurrent.duration.FiniteDuration