+++ src/test/scala/edu/ie3/simona/agent/grid/powerflow/DBFSAlgorithmFailedPowerFlowSpec.scala
@@ -93,7 +93,7 @@ class DBFSAlgorithmFailedPowerFlowSpec
   given GridAgentConstantData = GridAgentConstantData(
     gridAgentCoordinator.ref,
     environmentRefs,
-    simonaConfig.simona,
+    simonaConfig,
     3600,
     startTime,
     endTime,
@@ -110,12 +110,12 @@ class DBFSAlgorithmFailedPowerFlowSpec
         VoltageLimits(0.9, 1.1),
         startTime,
         endTime,
-        simonaConfig.simona,
+        simonaConfig,
       )
 
       val gridAgentInitData = GridAgentInitData(
         gridModel,
-        PowerFlowParams(simonaConfig.simona.powerflow.value),
+        PowerFlowParams(simonaConfig.powerflow.value),
       )
 
       val centerGridAgent = testKit.spawn(GridAgent(gridAgentInitData))
@@ -363,12 +363,12 @@ class DBFSAlgorithmFailedPowerFlowSpec
         VoltageLimits(0.9, 1.1),
         startTime,
         endTime,
-        simonaConfig.simona,
+        simonaConfig,
       )
 
       val gridAgentInitData = GridAgentInitData(
         gridModel,
-        PowerFlowParams(simonaConfig.simona.powerflow.value),
+        PowerFlowParams(simonaConfig.powerflow.value),
       )
 
       val slackGridAgent = testKit.spawn(GridAgent(gridAgentInitData))